src/requirements.txt
blinker==1.6.2
click==8.1.7
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==3.0.0
python-dotenv==0.18.0
mysql-connector-python==8.1.0
gunicorn==21.2.0